Fix bug in apply_rotary_pos_emb_flashatt: in Qwen2-5-VL (#36065)

+++ src/transformers/models/qwen2_5_vl/modular_qwen2_5_vl.py
@@ -65,8 +65,8 @@ else:
 
 def apply_rotary_pos_emb_flashatt(tensor: torch.Tensor, freqs: torch.Tensor) -> torch.Tensor:
     tensor_ = tensor.float()
-    cos = freqs.cos()
-    sin = freqs.sin()
+    cos = freqs.cos().float()
+    sin = freqs.sin().float()
     output = apply_rotary_emb(tensor_, cos, sin).type_as(tensor)
     return output
